Put Args in a separate header + update README.
