+++ include/tetwild/tetwild.h
@@ -8,72 +8,13 @@
 //
 // Created by Yixin Hu on 5/31/18.
 //
-#ifndef TETWILD_TETWILD_H
-#define TETWILD_TETWILD_H
 
-#include <array>
-#include <vector>
-#include <string>
+#pragma once
+
+#include <tetwild/Args.h>
 #include <Eigen/Dense>
 
 namespace tetwild {
-
-// Global arguments controlling the behavior of TetWild
-struct Args {
-    // Initial target edge-length at every vertex (in % of the bbox diagonal)
-    double initial_edge_len_rel = 5.0;
-
-    // Target epsilon (in % of the bbox diagonal)
-    double eps_rel = 0.1;
-
-    //////////////////////
-    // Advanced options //
-    //////////////////////
-
-    // Explicitly specify a sampling distance for triangles (in % of the bbox diagonal)
-    int sampling_dist_rel = -1;
-
-    // Run the algorithm in stage (as explain in p.8 of the paper)
-    // If the first stage didn't succeed, call again with `stage = 2`,  etc.
-    int stage = 1;
-
-    // Multiplier for resizing the target-edge length around bad-quality vertices
-    // See MeshRefinement::updateScalarField() for more details
-    double adaptive_scalar = 0.6;
-
-    // Energy threshold
-    // If the max tet energy is below this threshold, the mesh optimization process is stopped.
-    // Also used to determine where to resize the scalar field (if a tet incident to a vertex has larger energy than this threshold, then resize around this vertex).
-    double filter_energy_thres = 10;
-
-    // Threshold on the energy delta (avg and max) below which to rescale the target edge length scalar field
-    double delta_energy_thres = 0.1;
-
-    // Maximum number of mesh optimization iterations
-    int max_num_passes = 80;
-
-    // Sample points at voxel centers for initial Delaunay triangulation
-    bool use_voxel_stuffing = true;
-
-    // Use Laplacian smoothing on the faces/vertices covering an open boundary after the mesh optimization step (post-processing)
-    bool smooth_open_boundary = false;
-
-    // Target number of vertices (minimum), within 5% of tolerance
-    int target_num_vertices = -1;
-
-    // Background mesh for the edge length sizing field
-    std::string background_mesh = "";
-
-    // [debug] logging
-    bool write_csv_file = true;
-    std::string working_dir = "";
-    std::string postfix = "_";
-    std::string csv_file = "";
-    int save_mid_result = -1; // save intermediate result
-
-    bool is_quiet = false;
-};
-
 
 ///
 /// Robust tetrahedralization of an input triangle soup, with an envelop constraint.
@@ -82,7 +23,8 @@ struct Args {
 /// @param[in]  FI    { #FI x 3 input mesh triangles }
 /// @param[out] VO    { #VO x 3 output mesh vertices }
 /// @param[out] TO    { #TO x 4 output mesh tetrahedra }
-/// @param[out] A     { #TO x 1 array of min dihedral angle over each tet }
+/// @param[out] AO    { #TO x 1 array of min dihedral angle over each tet }
+/// @param[in]  args  { Extra arguments controlling the behavior of TetWild }
 ///
 void tetrahedralization(const Eigen::MatrixXd &VI, const Eigen::MatrixXi &FI,
     Eigen::MatrixXd &VO, Eigen::MatrixXi &TO, Eigen::VectorXd &AO, const Args &args = Args());
@@ -98,6 +40,5 @@ void tetrahedralization(const Eigen::MatrixXd &VI, const Eigen::MatrixXi &FI,
 void extractSurfaceMesh(const Eigen::MatrixXd &VI, const Eigen::MatrixXi &TI,
     Eigen::MatrixXd &VS, Eigen::MatrixXi &FS);
 
-}
+} // namespace tetwild
 
-#endif //TETWILD_TETWILD_H
